+++ modules/lang-painless/src/main/java/org/elasticsearch/painless/PainlessPlugin.java
@@ -38,11 +38,13 @@ import org.elasticsearch.repositories.RepositoriesService;
 import org.elasticsearch.rest.RestController;
 import org.elasticsearch.rest.RestHandler;
 import org.elasticsearch.script.IngestScript;
+import org.elasticsearch.script.NumberSortScript;
 import org.elasticsearch.script.ScoreScript;
 import org.elasticsearch.script.ScriptContext;
 import org.elasticsearch.script.ScriptEngine;
 import org.elasticsearch.script.ScriptModule;
 import org.elasticsearch.script.ScriptService;
+import org.elasticsearch.script.StringSortScript;
 import org.elasticsearch.search.aggregations.pipeline.MovingFunctionScript;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.watcher.ResourceWatcherService;
@@ -80,7 +82,9 @@ public final class PainlessPlugin extends Plugin implements ScriptPlugin, Extens
         // Functions used for scoring docs
         List<Whitelist> scoreFn = new ArrayList<>(Whitelist.BASE_WHITELISTS);
         Whitelist scoreFnWhitelist = WhitelistLoader.loadFromResourceFiles(Whitelist.class, "org.elasticsearch.score.txt");
+        Whitelist scoreFieldWhitelist = WhitelistLoader.loadFromResourceFiles(Whitelist.class, "org.elasticsearch.script.fields.score.txt");
         scoreFn.add(scoreFnWhitelist);
+        scoreFn.add(scoreFieldWhitelist);
         map.put(ScoreScript.CONTEXT, scoreFn);
 
         // Functions available to ingest pipelines
@@ -94,6 +98,16 @@ public final class PainlessPlugin extends Plugin implements ScriptPlugin, Extens
         for (ScriptContext<?> scriptContext : ScriptModule.RUNTIME_FIELDS_CONTEXTS) {
             map.put(scriptContext, getRuntimeFieldWhitelist(scriptContext.name));
         }
+
+        List<Whitelist> numSort = new ArrayList<>(Whitelist.BASE_WHITELISTS);
+        Whitelist numSortField = WhitelistLoader.loadFromResourceFiles(Whitelist.class, "org.elasticsearch.script.fields.numbersort.txt");
+        numSort.add(numSortField);
+        map.put(NumberSortScript.CONTEXT, numSort);
+
+        List<Whitelist> strSort = new ArrayList<>(Whitelist.BASE_WHITELISTS);
+        Whitelist strSortField = WhitelistLoader.loadFromResourceFiles(Whitelist.class, "org.elasticsearch.script.fields.stringsort.txt");
+        strSort.add(strSortField);
+        map.put(StringSortScript.CONTEXT, strSort);
 
         // Execute context gets everything
         List<Whitelist> test = new ArrayList<>(Whitelist.BASE_WHITELISTS);
