+++ server/src/main/java/org/elasticsearch/script/StringSortScript.java
@@ -37,4 +37,16 @@ public abstract class StringSortScript extends AbstractSortScript {
     public interface Factory extends ScriptFactory {
         LeafFactory newFactory(Map<String, Object> params, SearchLookup lookup);
     }
+
+    public static class FieldAccess {
+        private final StringSortScript script;
+
+        public FieldAccess(StringSortScript script) {
+            this.script = script;
+        }
+
+        public Field<?> field(String fieldName) {
+            return script.field(fieldName);
+        }
+    }
 }
